move sstructure and similarity searches to its own module
