+++ src/glados/models.py
@@ -184,6 +184,8 @@ class SSSearchJob(models.Model):
     SUBSTRUCTURE = 'SUBSTRUCTURE'
     CONNECTIVITY = 'CONNECTIVITY'
 
+    CONTEXT_PREFIX = '_context'
+
     SEARCH_TYPES = (
         (SIMILARITY, SIMILARITY),
         (SUBSTRUCTURE, SUBSTRUCTURE),
