+++ src/glados/static/coffee/Settings.coffee
@@ -303,8 +303,8 @@ glados.loadSearchResultsURLS = ()->
   glados.Settings.ELASTICSEARCH_CACHE = 'elasticsearch_cache'
   glados.Settings.REGISTER_USAGE_ENDPOINT = 'register_usage'
   glados.Settings.REGISTER_SEARCH_ENDPOINT = 'register_search'
-  glados.Settings.CHEMBL_LIST_HELPER_ENDPOINT = 'chembl_list_helper'
-  glados.Settings.CHEMBL_SUBMIT_STRUCTURE_SEARCH_ENDPOINT = 'submit_sustructure_search'
+  glados.Settings.CHEMBL_LIST_HELPER_ENDPOINT = 'api/chembl/list_pagination/get_page'
+  glados.Settings.CHEMBL_SUBMIT_STRUCTURE_SEARCH_ENDPOINT = 'api/chembl/sssearch/submit_substructure_search/'
   glados.Settings.CHEMBL_LIST_HELPER_URL = "#{glados.Settings.GLADOS_BASE_PATH_REL}#{glados.Settings.CHEMBL_LIST_HELPER_ENDPOINT}"
   glados.Settings.SHORTEN_URLS_URL = "#{glados.Settings.GLADOS_BASE_PATH_REL}#{glados.Settings.SHORTEN_URLS_ENDPOINT}"
   glados.Settings.EXTEND_URLS_ENDPOINT_URL = "#{glados.Settings.GLADOS_BASE_PATH_REL}#{glados.Settings.EXTEND_URLS_ENDPOINT}"
