+++ src/glados/views.py
@@ -15,8 +15,6 @@ from . import og_tags_generator
 from . import schema_tags_generator
 from . import glados_server_statistics
 from . import heatmap_helper
-from . import structure_and_sequence_searches_helper
-import traceback
 
 
 def visualise(request):
@@ -405,71 +403,6 @@ def elasticsearch_cache(request):
         raw_search_data = request.POST.get('search_data', '')
 
         response = glados_server_statistics.get_and_record_es_cached_response(index_name, raw_search_data)
-        if response is None:
-            return HttpResponse('ELASTIC SEARCH RESPONSE IS EMPTY!', status=500)
-
-        return JsonResponse(response)
-    else:
-        return JsonResponse({'error': 'this is only available via POST! You crazy hacker! :P'})
-
-
-def submit_sustructure_search(request):
-
-    if request.method == "POST":
-
-        search_type = request.POST.get('search_type')
-        raw_search_params = request.POST.get('raw_search_params')
-
-        try:
-
-            response = structure_and_sequence_searches_helper.generate_search_job(search_type, raw_search_params)
-            return JsonResponse(response)
-
-        except Exception as e:
-
-            traceback.print_exc()
-            return HttpResponse("Internal Server Error: {}".format(repr(e)), status=500)
-
-    else:
-        return JsonResponse({'error': 'this is only available via POST! You crazy hacker! :P'})
-
-
-def get_sssearch_status(request, search_id):
-
-    if request.method != "GET":
-        return JsonResponse({'error': 'This is only available via GET'})
-
-    try:
-        response = structure_and_sequence_searches_helper.get_sssearch_status(search_id)
-        return JsonResponse(response)
-    except Exception as e:
-        traceback.print_exc()
-        return HttpResponse('Internal Server Error', status=500)
-
-
-def chembl_list_helper(request):
-
-    if request.method == "POST":
-
-        index_name = request.POST.get('index_name', '')
-        raw_search_data = request.POST.get('search_data', '')
-        context_id = request.POST.get('context_id')
-        id_property = request.POST.get('id_property')
-        raw_contextual_sort_data = request.POST.get('contextual_sort_data')
-        print('raw_search_data: ', raw_search_data)
-
-        try:
-            if context_id is None or context_id == 'undefined' or context_id == 'null':
-                print('no context id')
-                response = glados_server_statistics.get_and_record_es_cached_response(index_name, raw_search_data)
-            else:
-                response = structure_and_sequence_searches_helper.get_items_with_context(index_name, raw_search_data,
-                                                                                         context_id, id_property,
-                                                                                         raw_contextual_sort_data)
-        except Exception as e:
-            traceback.print_exc()
-            return HttpResponse('Internal Server Error', status=500)
-
         if response is None:
             return HttpResponse('ELASTIC SEARCH RESPONSE IS EMPTY!', status=500)
